Use pdb_addr2line::TypeFormatter.
Fixes #279.
