+++ Cargo.toml
@@ -32,7 +32,7 @@ hashbrown = { version = "0.11", features = ["serde"] }
 lazy_static = "1.4"
 log = "0.4"
 num_cpus = "1.13"
-pdb = "0.7"
+pdb-addr2line = "0.7"
 regex = "1.4"
 reqwest = { version = "0.11", optional = true, default-features = false, features = [
     "blocking",
