+++ src/windows/lines.rs
@@ -6,6 +6,7 @@
 use crate::common::LineFinalizer;
 use crate::line::{Line, Lines};
 use pdb::{AddressMap, PdbInternalRva};
+use pdb_addr2line::pdb;
 
 impl LineFinalizer<AddressMap<'_>> for Lines {
     fn finalize(&mut self, sym_rva: u32, sym_len: u32, address_map: &AddressMap) {
