+++ src/windows/mod.rs
@@ -7,5 +7,4 @@ mod lines;
 pub mod pdb;
 mod source;
 mod symbol;
-mod types;
 pub mod utils;
